% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addIslands.sp_pkg.R
\name{addIslands.sp}
\alias{addIslands.sp}
\title{Adding new islands to global islands dataset}
\usage{
addIslands.sp(newIslands, currentIslands, polys)
}
\arguments{
\item{newIslands}{Data frame of new islands, with a row for each new island. Must have columns called 'lat' and 'long' but no other requirements.}

\item{currentIslands}{SpatialPolygonsDataFrame of current global island dataset. MUST have numeric ID column in attributes named 'uniqueID'. Unprojected lat/long.}

\item{polys}{SpatialPolygonsDataFrame of all land on earth (GADM shapefile, for example). Unprojected lat/long.}
}
\value{
Returns spatialPolygonsDataFrame of newIslands. This SPDF can be merged with currentIslands easily using rbind if the columns are the same...see newCurrentIslands <- rbind(currentIslands, polys2add, makeUniqueIDs = T)
}
\description{
This function adds new islands to the global island dataset, and assigns unique numeric IDs to the new islands based off the ID system of the old islands.
}
\author{
Tyler J Tran, \email{tylerjtran@gmail.com}
}
