% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindsheet.R
\name{bindsheet}
\alias{bindsheet}
\alias{bindsheet.morph}
\title{General functions for googlesheet data manipulation}
\usage{
bindsheet.morph(sheets)
}
\arguments{
\item{sheets}{character vector of sheet titles to bind}
}
\value{
concatenated datatable
}
\description{
concatenates sheets of similar data
}
\details{
bindsheet.morph assumes that sheets to bind have the same columns
}
\author{
Matthew R. Helmus
}

