% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIslandArea_pkg.R
\name{calcIslandArea}
\alias{calcIslandArea}
\title{Calculating island area}
\usage{
calcIslandArea(polys)
}
\arguments{
\item{polys}{SpatialPolygonsDataFrame of your global islands dataset. Unprojected lat/long.}
}
\value{
Returns vector of polygon areas. Can be added as a field in attribute table of SPDF of global islands.
}
\description{
This function re-projects a spatialPolygonsDataFrame (from shapefile) of islands in the global islands dataset to the World cylindrical equal area projection, and calculates area for each island (or each polygon, rather).
}
\author{
Tyler J Tran, \email{tylerjtran@gmail.com}
}
