% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/distance.reproj_pkg.R
\name{distance.reproj}
\alias{distance.reproj}
\title{Find distances between island centroids and other polygons}
\usage{
distance.reproj(pts, polys)
}
\arguments{
\item{pts}{SpatialPointsDataFrame of island centroids. Assumed unprojected lat/long.}

\item{polys}{SpatialPolygonsDataFrame of islands for which you want to calculate distances between centroids and polys. Assumed unprojected lat/long.}
}
\value{
Returns a vector of distances between each of the centroids and the closest island. The vector is in the same order as the centroid input.
}
\description{
This function calculates the distance between island centroids (input) and the nearest polygons. The function iterates through however many centroid points are given, and for each point, reprojects to an azimuthal equidistant projection that is centered around the centroid (see Weigelt and Kreft 2013). Because of the use of a for loop and reprojections, this can be a very slow script.
}
\author{
Tyler J Tran, \email{tylerjtran@gmail.com}
}
