% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findCentroids_pkg.R
\name{findCentroids}
\alias{findCentroids}
\title{Find centroids of island polygons (global islands)}
\usage{
findCentroids(polys)
}
\arguments{
\item{polys}{SpatialPolygonsDataFrame of islands for which you want centroids. Assumed unprojected lat/long.}
}
\value{
Returns data frame of lat/long coordinates of the centroids for each island.
}
\description{
This function takes a polygon shapefile as input and finds the centroids. These can then be used to calculate geographic isolation (distances between island centroids and other islands, continents, etc.)
}
\author{
Tyler J Tran, \email{tylerjtran@gmail.com}
}
