% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/igrade.R
\name{igrade}
\alias{igrade}
\alias{do.grades.ieco}
\alias{getlet.ieco}
\title{Calculate distribution of grades}
\usage{
do.grades.ieco(grades, divs = NULL, cut = 0, tit = "Exam Grades",
  breaks = length(grades)/3, labels = c("F", "D", "C", "B", "A"), ...)

getlet.ieco(grades, divs, full.divs = c(0, 60, 63, 67, 70, 73, 77, 80, 83, 87,
  90, 93, 97), full.labels = c("F", "D-", "D", "D+", "C-", "C", "C+", "B-",
  "B", "B+", "A-", "A", "A+"), noAplus = TRUE)
}
\arguments{
\item{grades}{Raw grades}

\item{divs}{divisions for grades (optional)}

\item{cut}{low end Cut off to remove 0 from statistics}

\item{tit}{Title for Figure}

\item{breaks}{breaks for the histogram, default=length(grades)/3}

\item{labels}{grade labels, default are the U.S.A. system of five letter grades, ABCDF}

\item{\dots}{other parameters for hist}

\item{full.divs}{divisions for all grades not just main grades as in div}

\item{full.labels}{names of all the grades}

\item{noAplus}{remove the first (A+) listed grade and assign the second (A)}

\item{linear}{default assignment of subgrades by interpolation}
}
\value{
the data set
}
\description{
Calculate the grades of a class of students, given raw scores on exam
}
\details{
To remove students who do not take the test a low end cut off is used to excise any grades below that level.  Both mean, and standard deviations are shown as well as median and quartiles.
}
\examples{
g = rnorm(n=130, m=82, sd=10)
g[g>100] = 100
g[g<1] = 1

B = boxplot(g)
##########   set divisions automatically:
divs = c(min(g), B$stats[1:4] + diff(B$stats)/2, max(g) )
D1 = do.grades.ieco(g, divs=divs, tit="GEOL 105 Exam 1")
}
\author{
Jonathan M. Lees from the ProfessR package, with edits by M.R. Helmus
}
