% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nearestLarger_pkg.R
\name{nearestLarger}
\alias{nearestLarger}
\title{Calculating distance to nearest island of equal or larger area}
\usage{
nearestLarger(pts, polys)
}
\arguments{
\item{pts}{SpatialPointsDataFrame of centroids of your islands. Can use findCentroids() function.}

\item{polys}{SpatialPolygonsDataFrame of islands in the world. Example GADM.}
}
\value{
Returns vector of distances between each of the centroids and the nearest island of equal or larger area.
}
\description{
This function calculates distances, for each island, to the nearest island of equal or larger area. It calculates area using the calcIslandArea() function from this package, and then reprojects into the azimuthal equidistant projection with the projection centered on the centroid for each iteration.
}
\author{
Tyler J Tran, \email{tylerjtran@gmail.com}
}
