% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phylo_compare.R
\name{phylo_compare}
\alias{phylo_compare}
\alias{compare_multiPhylo}
\title{Compare the structure of phylogenies}
\usage{
compare_multiPhylo(multitree, which.phylo = 1, ...)
}
\arguments{
\item{multitree}{list of phylogenies of class \code{multiPhylo}}

\item{which.phylo}{index location of the phylogeny with multitree to compare, if NULL then all trees are compared to each other}

\item{...}{further arguments to be passed to \code{\link[ape]{all.equal.phylo}}}
}
\value{
a data.frame of comparisons
}
\description{
These functions are wrappers of existing functions that makes it easy to compare and then summarize the toplogy, structure and ages of phylogenies.

\code{compare_multiPhylo}
}
\seealso{
\code{\link[RADami]{compare.all.trees}} \code{\link[ape]{all.equal.phylo}}
}
\author{
Matthew R. Helmus inspired by code from Andrew L. Hipp in \code{\link[RADami]{compare.all.trees}}
}
