% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scraping.R
\name{web_scrape}
\alias{web_scrape}
\alias{google.counts}
\title{Calculate distribution of grades}
\usage{
google.counts(s)
}
\arguments{
\item{s}{a search term}
}
\value{
numeric
}
\description{
various functions to scrape data from the web
}
\details{
Number of approximate google search hits
}
\examples{
google.counts("iecolab")
}
\author{
Drew Conway, with edits by M.R. Helmus, https://gist.github.com/drewconway/791559
}
