% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindsheet.R
\name{sheet.data}
\alias{bindsheet.morph}
\alias{readsheet.morph}
\alias{sheet.data}
\title{General functions for googlesheet data manipulation}
\usage{
readsheet.morph(tle)

bindsheet.morph(sheets)
}
\arguments{
\item{sheets}{character vector of sheet titles to bind}

\item{title}{character object of the title of the sheet to load}
}
\value{
concatenated datatable
}
\description{
concatenates sheets of similar data
}
\details{
bindsheet.morph assumes that sheets to bind have the same columns
}
\author{
Matthew R. Helmus
}

