% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sppList_pkg.R
\name{sppList}
\alias{sppList}
\title{Get a list of species for islands}
\usage{
sppList(islands, sppRanges, IDvar)
}
\arguments{
\item{islands}{SpatialPolygonsDataFrame of islands}

\item{sppRanges}{SpatialPolygonsDataFrame of species ranges}

\item{IDvar}{A character string. The name of the column in shapefile sppRanges with the species name to be output in the list.}
}
\value{
Returns a list, with one entry for each island, of the species ranges that overlap for that island.
}
\description{
This function takes species range maps and a shapefile of islands and outputs a list of species on each island. A few notes: if having problems with orphaned holes, try readShapePoly instead of readOGR. Also, if the output list is not the same length as the number of islands, you might need to manually add elements to the end of the list that are NA. The last element of the output list will likely be an island that is not NA (i.e. an island with species ranges that overlap).
}
\author{
Tyler J Tran, \email{tylerjtran@gmail.com}
}
