% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tip_grafts.R
\name{tip_grafts}
\alias{tip_grafts}
\alias{tipXtip}
\title{Grafts tips on phylogenies}
\usage{
tipXtip(tree, addtip, where.tip = NULL)
}
\arguments{
\item{tree}{phylgeny of class \code{\link[ape]{phylo}}}

\item{addtip}{name of the tip to add}

\item{where.tip}{name of the tip}
}
\value{
a \code{\link[ape]{phylo}} with the grafted tip
}
\description{
Places tips where you want a tip to go on a phylogeny. These functions are good for adding missing taxa in ecophylogenetic analyses when you have a good phylogeny, and only need to graft a few missing taxa with a good hypothesis as to where they should go.
\code{tipXtip} adds a tip halfway the distance between a supplied tip and its MRCA.
}
\seealso{
\code{\link[phytools]{bind.tip}} \code{\link[ape]{bind.tree}}
}
\author{
Matthew R. Helmus
}
