% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait.check.R
\name{trait.check}
\alias{dp.count}
\alias{dp.spp}
\alias{trait.check}
\title{Check the data for repeats and errors}
\usage{
dp.spp(x, ids = NULL)

dp.count(x, ids = NULL)
}
\arguments{
\item{x}{Dataset to be parsed}

\item{ids}{specimen id numbers that you want to check otherwise all in x}
}
\value{
the value of the identified statistic
}
\description{
basic checking functions for morphology data to see if ids match spp, misspellings etc.
}
\details{
NONE
}
\author{
Matthew R. Helmus
}
\references{
None None
}
\seealso{
None None
}

