% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait.new.R
\name{trait.new}
\alias{svl_trait}
\alias{tail_apply}
\alias{tail_total}
\alias{trait.new}
\title{Add and edit traits to a global traits data set}
\usage{
tail_total(num, x, newnumber = 9.1, newtrait = "tail total length")

tail_apply(ids, x, newnumber = 9.1, newtrait = "tail total length")

svl_trait(x, counts = NULL, these = c(2, 3, 4, 5, 6, 7, 9, 10, 11, 12, 13,
  14, 15, 15.2, 16.2, 17, 18, 19.3, 19.4, 20, 21, 22.2, 23, 24, 25.3, 25.4, 26,
  27))
}
\arguments{
\item{num}{specimen number}

\item{x}{Dataset that traits should be added}

\item{newnumber}{Number of the new trait}

\item{newtrait}{Name of the trait that you want to add}

\item{ids}{vector of specimen ids}

\item{counts}{unique of the counts column, vector of the unique id for each time a specimen was measured}

\item{these}{trait numbers to correct for svl}
}
\value{
the data set with the appended traits
}
\description{
combine or regress individual traits
}
\details{
for an individual measured specimen these functions combine traits
}
\author{
Matthew R. Helmus
}

