% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait.stats.R
\name{trat.stat}
\alias{trait.mean}
\alias{trait.n}
\alias{trait.range}
\alias{trait.sd}
\alias{trat.stat}
\title{Statistics to summarize Global Traits}
\usage{
trait.mean(num, x)

trait.n(num, x)

trait.sd(num, x)

trait.range(num, x)
}
\arguments{
\item{num}{specimen id number}

\item{x}{Dataset to be parsed}
}
\value{
the value of the identified statistic
}
\description{
Means, ranges etc. for each individual specimens measured.
}
\details{
best when used with apply across a species
}
\author{
Matthew R. Helmus
}

