% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/unep_pkg.R
\name{unep}
\alias{unep}
\title{UNEP Isolation Index}
\usage{
unep(nearestLarger, nearestArch, nearestContinent)
}
\arguments{
\item{nearestLarger}{Numeric object or vector indicating the distance between the island of interest and the nearest island of equivalent or larger size}

\item{nearestArch}{Numeric object or vector for distance to nearest archipelago or island group}

\item{nearestContinent}{Numeric object or vector for didstance to nearest continent}
}
\description{
This function calculates the UNEP isolation index for islands
}
\author{
Tyler J Tran, \email{tylerjtran@gmail.com}
}
