% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vif_Lmer.R
\name{vif_lmer}
\alias{vif_lmer}
\alias{vif.mer}
\title{VIF for lmer fit}
\usage{
vif.mer(fit)
}
\arguments{
\item{fit}{lmer fit from \code{\link{lmer}}}
}
\value{
a vector of VIF values
}
\description{
calculates VIF for the fixed effects in lmer fits
}
\seealso{
\url{https://jonlefcheck.net/2012/12/28/dealing-with-multicollinearity-using-variance-inflation-factors/}
}
\author{
Jon Lefcheck see website
}
