% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bindsheet.R
\name{sheet.data}
\alias{sheet.data}
\alias{readsheet}
\alias{bindsheet.morph}
\title{General functions for googlesheet data manipulation}
\usage{
readsheet(tle)

bindsheet.morph(sheets, fnadd = TRUE)
}
\arguments{
\item{tle}{character object of the title of the sheet to load}

\item{sheets}{character vector of sheet titles to bind}

\item{fnadd}{add the title of the sheets as a column in the bind}
}
\value{
concatenated datatable
}
\description{
concatenates sheets of similar data
}
\details{
bindsheet.morph assumes that sheets to bind have the same columns and is only for herp morphology datasheets
}
\author{
Matthew R. Helmus test
}
