% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tip_grafts.R
\name{tip_grafts}
\alias{tip_grafts}
\alias{tipXtip}
\alias{tipXnode}
\alias{treeXnode}
\alias{tree_scale}
\title{Grafts tips on phylogenies}
\usage{
tipXtip(tree, addtip, where.tip = NULL)

tipXnode(tree, addtip, where.nodes = NULL)

treeXnode(tree, addtree, where.nodes = NULL, tip.out = "tip.out")

tree_scale(tree, scale)
}
\arguments{
\item{tree}{phylgeny of class \code{\link[ape]{phylo}}, this is the starting tree}

\item{addtip}{name of the tip to add}

\item{where.tip}{name of the tip}

\item{where.nodes}{vector of two node numbers to insert the tip between}

\item{addtree}{phylgeny of class \code{\link[ape]{phylo}} to add}

\item{tip.out}{tip label of the outgroup tip that defines the stem of the clade to be grafted}

\item{scale}{scalar to multiply all edges of a tree}
}
\value{
a \code{\link[ape]{phylo}} with the grafted tip
}
\description{
Places tips and clades where you want on a phylogeny. These functions are good for adding missing taxa in ecophylogenetic analyses when you have a good phylogeny, and only need to graft a few missing taxa with a good hypothesis as to where they should go.

\code{tipXtip} adds a tip halfway the distance between a supplied tip and its MRCA.

\code{tipXnode} adds a tip halfway the distance between a supplied set of nodes that define one edge

\code{treeXnode} adds a clade halfway the distance between a supplied set of nodes that define one edge

\code{tree_scale} linear scaling of a tree, use this to change the root age of a tree to the \code{scale} parameter, while scaling all branch lengths and node ages proportional to this change. Does not change relative relationships among nodes and tips.
}
\seealso{
\code{\link[phytools]{bind.tip}} \code{\link[ape]{bind.tree}}

for \code{tree_scale} see the code at \href{http://blog.phytools.org/2012/02/quicker-way-to-rescale-total-length-of.html}{Phytools Blog} and \code{\link[geiger]{rescaleTree}}
}
\author{
Matthew R. Helmus
}
