% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait.checks.R
<<<<<<< HEAD
\name{dp.spp}
\alias{dp.spp}
=======
\name{trat.check}
\alias{dp.spp}
\alias{trat.check}
>>>>>>> efcbc2e6ea221da864829252212bba9b92fe1d38
\title{Statistics to summarize Global Traits}
\usage{
dp.spp(x, ids = NULL)
}
\arguments{
\item{x}{Dataset to be parsed}

\item{ids}{specimen id numbers that you want to check otherwise all in x}
}
\value{
the value of the identified statistic
}
\description{
basic checking functions for morphology data to see if ids match spp, misspellings etc.
}
\author{
Matthew R. Helmus
}

