% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trait.stats.R
\name{trat.stat}
\alias{trat.stat}
\alias{trait.mean}
\alias{trait.n}
\alias{trait.sd}
\alias{trait.range}
\alias{specimen.mean}
\alias{specimen.n}
\alias{specimen.sd}
\alias{specimen.range}
\alias{spp.mean}
\alias{spp.n}
\alias{spp.sd}
\alias{spp.range}
\title{Statistics to summarize Global Traits}
\usage{
trait.mean(num, x)

trait.n(num, x)

trait.sd(num, x)

trait.range(num, x)

specimen.mean(x, ids)

specimen.n(x, ids)

specimen.sd(x, ids)

specimen.range(x, ids)

spp.mean(x, ids)

spp.n(x, ids)

spp.sd(x, ids)

spp.range(x, ids)
}
\arguments{
\item{num}{specimen id number}

\item{x}{Dataset to be parsed}

\item{ids}{Vector of specimen id numbers}
}
\value{
the value of the identified statistic
}
\description{
Means, ranges etc. for each individual specimens measured.
}
\details{
The spp functions build on the specimen functions which build on the trait functions. If you just want a data set of traits aggregated to species, use the spp functions
}
\author{
Matthew R. Helmus
}
